src/modulo-01-java-OO/projeto-lotr-bluej/DataTerceiraEra.java
public class DataTerceiraEra {
    private int dia, mes, ano;
    
    public DataTerceiraEra(int dia, int mes, int ano) {
        this.dia = dia;
        this.mes = mes;
        this.ano = ano;
    }
    
    public int getDia() {
        return this.dia;
    }
    
    public int getMes() {
        return this.mes;
    }
    
    public int getAno() {
        return this.ano;
    }
    
    public boolean ehBissexto() {
        // São bissextos todos os anos múltiplos de 400, p.ex: 1600, 2000, 2400, 2800...
        return this.ano % 400 == 0 || this.ano % 4 == 0 && this.ano % 100 != 0;
    }
    
    // Sobrescrita do método equals para comparar todas propriedades, e não só as referências  
    public boolean equals(DataTerceiraEra data){ 
        return this.dia == data.getDia() 
            && this.mes == data.getMes() 
            && this.ano == data.getAno();
    }
}